svlog: codegen: Emit void funcs with no args
Emit the most basic of functions, and associated calls. Currently the
LLHD crate doesn't allow for the void return value of a function to be
used, which prevents code generation. This needs to be sorted out in
LLHD first (e.g. add `const void` and make `call()` always return a
value).